refactor(server): split auth route handlers
