+++ nodelite-server/src/handlers/mod.rs
@@ -2,7 +2,7 @@
 //!
 //! 这里的 `mod.rs` 只负责拼装子模块导出,把 handler 按职责拆到更窄的文件里:
 //! - `pages`: HTML 页面与静态 UI 资源;
-//! - `auth_routes`: 只读认证、2FA 校验与健康探针;
+//! - `auth_routes`: 只读认证、2FA 校验与健康探针拼装;
 //! - `api`: 仪表盘读取接口与 Prometheus 导出;
 //! - `install`: Agent 安装脚本与 bootstrap 下发;
 //! - `settings`: 管理面板的写操作与设置查询。
